requirements.txt
streamlit
Pillow
torch
numpy
scikit-learn
git+https://github.com/openai/CLIP.git
